modify methods with mid-sized structs as args to pass by const ref instead of by value
