+++ include/game_board.hpp
@@ -26,16 +26,16 @@ public:
   GameBoard();
   GameBoard(const BoardMapInt_t starting_board);
   vector<BoardSpace> ImplementGetAllSpacesOccupiedBy(PieceColor color) const;
-  PieceColor ImplementGetColor(BoardSpace space) const;
-  PieceType ImplementGetType(BoardSpace space) const;
+  PieceColor ImplementGetColor(const BoardSpace &space) const;
+  PieceType ImplementGetType(const BoardSpace &space) const;
   MoveCollection ImplementCalcFinalMovesOf(PieceColor color);
   bool IsInCheck(PieceColor color);
-  ExecutedMove ImplementExecuteMove(Move move);
-  bool IsCaptureMove(ExecutedMove executed_move);
-  void ImplementUndoMove(ExecutedMove executed_move);
-  GamePiece GetOccupantAt(BoardSpace space) const;
+  ExecutedMove ImplementExecuteMove(const Move &move);
+  bool IsCaptureMove(const ExecutedMove &executed_move) const;
+  void ImplementUndoMove(const ExecutedMove &executed_move);
+  GamePiece GetOccupantAt(const BoardSpace &space) const;
   const BoardMap_t &map() const;
-  void ImplementAttachMoveCallback(const function<void(ExecutedMove)>& callback);
+  void ImplementAttachMoveCallback(const function<void(const ExecutedMove&)>& callback);
   bool ImplementIsDraw();
   const std::map<PieceColor, vector<ExecutedMove>>& move_log() const;
 
@@ -48,7 +48,7 @@ private:
 
   //! Stores functions that are called after any change in board config to keep
   //! boardstate::SingleZobristTracker objects updated.
-  vector<function<void(ExecutedMove)>> move_callbacks_;
+  vector<function<void(const ExecutedMove&)>> move_callbacks_;
 
   //! Vectors of all moves that have been executed (and not un-done) by each player.
   std::map<PieceColor, vector<ExecutedMove>> move_log_;
@@ -56,10 +56,10 @@ private:
   //! Number of moves executed since last time a piece was captured. 
   int moves_since_last_capture_; 
 
-  void UpdateStateTracker(ExecutedMove executed_move);
-  void SetOccupantAt(BoardSpace space, GamePiece piece);
-  void AddToMoveLog(ExecutedMove executed_move);
-  void RemoveFromMoveLog(ExecutedMove executed_move);
+  void UpdateStateTracker(const ExecutedMove &executed_move);
+  void SetOccupantAt(const BoardSpace &space, GamePiece piece);
+  void AddToMoveLog(const ExecutedMove &executed_move);
+  void RemoveFromMoveLog(const ExecutedMove &executed_move);
   bool ViolatesRepeatRule(PieceColor color);
 };
