+++ include/zobrist.hpp
@@ -1,0 +1,135 @@
+
+
+#pragma once
+
+#include <array>
+#include <vector>
+
+namespace boardstate {
+
+//! Container for all of the hash keys needed to run a boardstate::SingleZobristTracker.
+template <typename KeyType>
+class ZobristCalculator {
+public:
+  typedef array<array<KeyType, kNumFiles>, kNumRanks> PieceZarray_t;
+  typedef array<PieceZarray_t, kNumPieceTypeVals> TeamZarray_t;
+  typedef array<TeamZarray_t, 2> GameZarray_t;
+
+  ZobristCalculator(uint32_t seed)
+      : zarray_{}
+      , turn_key_{}
+      , board_state_{}
+      , seed_{seed} {
+    PseudoRandomKeyGenerator<KeyType> key_generator{seed};
+    turn_key_ = key_generator.GenerateKey();
+    zarray_ = CreateGameZarray(key_generator);
+  };
+
+  ZobristCalculator()
+      : ZobristCalculator(random_device{}()) {}
+
+  void FullBoardStateCalc(const BoardMap_t &board_map) {
+    board_state_ = 0;
+    for (size_t rank = 0; rank < kNumRanks; rank++) {
+      for (size_t file = 0; file < kNumFiles; file++) {
+        if (board_map[rank][file].piece_color != 0) {
+          board_state_ = board_state_ ^ GetHashValueAt(
+                                            board_map[rank][file].piece_color,
+                                            board_map[rank][file].piece_type,
+                                            BoardSpace{(int)rank, (int)file}
+                                        );
+        }
+      }
+    }
+  }
+
+  void UpdateBoardState(const ExecutedMove &executed_move) {
+    board_state_ = board_state_ ^ GetHashValueAt(
+                                      executed_move.moving_piece.piece_color,
+                                      executed_move.moving_piece.piece_type,
+                                      executed_move.spaces.start
+                                  );
+
+    // if capture piece, remove from board
+    if (executed_move.destination_piece.piece_color != PieceColor::kNul) {
+      board_state_ = board_state_ ^ GetHashValueAt(
+                                        executed_move.destination_piece.piece_color,
+                                        executed_move.destination_piece.piece_type,
+                                        executed_move.spaces.end
+                                    );
+    }
+
+    // moving piece to new space
+    board_state_ = board_state_ ^ GetHashValueAt(
+                                      executed_move.moving_piece.piece_color,
+                                      executed_move.moving_piece.piece_type,
+                                      executed_move.spaces.end
+                                  );
+
+    // change state now that its other player's turn
+    board_state_ = board_state_ ^ turn_key();
+  }
+
+  KeyType GetHashValueAt(PieceColor color, PieceType piece_type, BoardSpace space) {
+    return zarray_[GetZColorIndexOf(color)][piece_type][space.rank][space.file];
+  }
+
+  //! Collects all keys into a std::vector (typically used for entropy calc).
+  std::vector<KeyType> KeysVector() const {
+    std::vector<KeyType> keys_vector;
+
+    for (const auto &team : zarray) {
+      for (const auto &piece_zarray : team) {
+        for (const auto &rank : piece_zarray) {
+          for (KeyType key : rank) {
+            keys_vector.push_back(key);
+          }
+        }
+      }
+    }
+    keys_vector.push_back(turn_key);
+
+    return keys_vector;
+  }
+
+  KeyType board_state() { return board_state_; }
+  GameZarray_t zarray() { return zarray_; }
+  KeyType turn_key() { return turn_key_; }
+  uint32_t seed() { return seed_; }
+
+private:
+  GameZarray_t zarray_;
+  KeyType turn_key_;
+  uint32_t seed_;
+  KeyType board_state_;
+
+  static const GameZarray_t CreateGameZarray(
+      PseudoRandomKeyGenerator<KeyType> &key_generator
+  ) {
+    GameZarray_t game_zarray{};
+    for (auto color_idx = 0; color_idx < 2; color_idx++) {
+      for (auto piece_id = 1; piece_id < kNumPieceTypeVals; piece_id++) {
+        for (auto rank = 0; rank < kNumRanks; rank++) {
+          for (auto file = 0; file < kNumFiles; file++) {
+            game_zarray[color_idx][piece_id][rank][file] = key_generator.GenerateKey();
+          }
+        }
+      }
+    }
+    return game_zarray;
+  };
+};
+
+// template <typename KeyType>
+// class ZobristComponent {
+// public:
+//   explicit ZobristComponent(const std::vector<ZobristCalculator<KeyType>> &calculators)
+//       : calculators_(calculators) {}
+
+//   void UpdateBoardState(const ExecutedMove& executed_move)
+
+// private:
+//   std::vector<ZobristCalculator<KeyType>> calculators_;
+// };
+
+} // namespace boardstate
